+++ src/main/java/com/quasiris/qsf/pipeline/filter/elastic/Profiles.java
@@ -1,0 +1,9 @@
+package com.quasiris.qsf.pipeline.filter.elastic;
+
+public class Profiles {
+
+
+    public static String matchAll() {
+        return "classpath://com/quasiris/qsf/elastic/profiles/match-all-profile.json";
+    }
+}